Initial work on cross-platform compat
Created Platform class in common, this establishes utility functions to check if client or server, the logic itself is implemented in fabric & forge respectively
Created Client class for client-only code in common, this has init and keyPress functions.

keyPress is further specialized because of the methodology difference between fabric & forge. So a single keyPress function is created to be called by both loaders in their respective domains.